Ensemble Orange #15:
Completed pulling GameMonitor out of Game and into GameService.

Wrote "paired" tests against GameService using the real JDBC and the Fake repositories to ensure we're doing things like "save" correctly. This surfaced a bug in CardDecoder#decodeCards that was returning the collector toList() for cards, which is immutable.

Refactored to use the Consumer functional interface in the GameService.